+++ zio-http-example/src/main/scala/example/CliExamples.scala
@@ -1,0 +1,136 @@
+package example
+
+import zio._
+import zio.http._
+import zio.http.codec._
+import zio.http.endpoint._
+import zio.http.endpoint.cli._
+import zio.schema._
+
+object TestCliApp extends zio.cli.ZIOCliDefault {
+  import HttpCodec._
+
+  final case class User(
+    @description("The unique identifier of the User")
+    id: Int,
+    @description("The user's name")
+    name: String,
+    @description("The user's email")
+    email: Option[String],
+  )
+  object User {
+    implicit val schema = DeriveSchema.gen[User]
+  }
+  final case class Post(
+    @description("The unique identifier of the User")
+    userId: Int,
+    @description("The unique identifier of the Post")
+    postId: Int,
+    @description("The post's contents")
+    contents: String,
+  )
+  object Post {
+    implicit val schema = DeriveSchema.gen[Post]
+  }
+
+  val getUser =
+    Endpoint
+      .get("users" / int("userId") ?? Doc.p("The unique identifier of the user"))
+      .header(HeaderCodec.location ?? Doc.p("The user's location"))
+      .out[User] ?? Doc.p("Get a user by ID")
+
+  val getUserPosts =
+    Endpoint
+      .get(
+        "users" / int("userId") ?? Doc.p("The unique identifier of the user") / "posts" / int("postId") ?? Doc.p(
+          "The unique identifier of the post",
+        ),
+      )
+      .query(query("name") ?? Doc.p("The user's name"))
+      .out[List[Post]] ?? Doc.p("Get a user's posts by userId and postId")
+
+  val createUser =
+    Endpoint
+      .post("users")
+      .in[User]
+      .out[String] ?? Doc.p("Create a new user")
+
+  val cliApp =
+    HttpCliApp.fromEndpoints(
+      name = "users-mgmt",
+      version = "0.0.1",
+      summary = "Users management CLI",
+      footer = "Copyright 2023",
+      host = "localhost",
+      port = 8080,
+      endpoints = Chunk(getUser, getUserPosts, createUser),
+    )
+}
+
+object TestCliServer extends zio.ZIOAppDefault {
+  import HttpCodec._
+
+  final case class User(
+    @description("The unique identifier of the User")
+    id: Int,
+    @description("The user's name")
+    name: String,
+    @description("The user's email")
+    email: Option[String],
+  )
+  object User {
+    implicit val schema = DeriveSchema.gen[User]
+  }
+  final case class Post(
+    @description("The unique identifier of the User")
+    userId: Int,
+    @description("The unique identifier of the Post")
+    postId: Int,
+    @description("The post's contents")
+    contents: String,
+  )
+  object Post {
+    implicit val schema = DeriveSchema.gen[Post]
+  }
+
+  val getUser =
+    Endpoint
+      .get("users" / int("userId") ?? Doc.p("The unique identifier of the user"))
+      .header(HeaderCodec.location ?? Doc.p("The user's location"))
+      .out[User] ?? Doc.p("Get a user by ID")
+
+  val getUserRoute =
+    getUser.implement { case (id, _) =>
+      ZIO.succeed(User(id, "Juanito", Some("juanito@test.com")))
+    }
+
+  val getUserPosts =
+    Endpoint
+      .get(
+        "users" / int("userId") ?? Doc.p("The unique identifier of the user") / "posts" / int("postId") ?? Doc.p(
+          "The unique identifier of the post",
+        ),
+      )
+      .query(query("name") ?? Doc.p("The user's name"))
+      .out[List[Post]] ?? Doc.p("Get a user's posts by userId and postId")
+
+  val getUserPostsRoute =
+    getUserPosts.implement { case (userId, postId, name) =>
+      ZIO.succeed(List(Post(userId, postId, name)))
+    }
+
+  val createUser =
+    Endpoint
+      .post("users")
+      .in[User]
+      .out[String] ?? Doc.p("Create a new user")
+
+  val createUserRoute =
+    createUser.implement { user =>
+      ZIO.succeed(user.name)
+    }
+
+  val routes = getUserRoute ++ getUserPostsRoute ++ createUserRoute
+
+  val run = Server.serve(routes.toApp).provide(Server.default)
+}
